fix: not install issue

+++ docker/install.sh
@@ -5,16 +5,19 @@ sudo apt remove -y --purge \
     docker-engine \
     docker.io \
     containerd \
-    runc \
-&& sudo apt autoremove -y --purge \
+    runc
+
+sudo apt autoremove -y --purge \
 && sudo apt update \
 && sudo apt install -y --no-install-recommends \
     ca-certificates \
     curl \
     gnupg \
     lsb-release \
-&& sudo mkdir -m 0755 -p /etc/apt/keyrings \
-&& curl -fsSL https://download.docker.com/linux/ubuntu/gpd \
+    pciutils \
+    kmod \
+&& sudo install -m 0755 -d /etc/apt/keyrings \
+&& curl -fsSL https://download.docker.com/linux/ubuntu/gpg \
     | sudo gpg --dearmor -o /etc/apt/keyrings/docker.gpg \
 && echo \
     "deb [arch=$(dpkg --print-architecture) signed-by=/etc/apt/keyrings/docker.gpg] https://download.docker.com/linux/ubuntu \
@@ -36,5 +39,24 @@ sudo apt remove -y --purge \
 && sudo systemctl enable docker.service \
 && sudo systemctl enable containerd.service
 
-
+if (lspci | grep -q VGA ||
+    lspci | grep -iq NVIDIA ||
+    lsmod | grep -q nvidia ||
+    nvidia-smi -L >/dev/null 2>&1 | grep -iq nvidia) &&
+    (command -v nvidia-smi >/dev/null 2>&1); then
+    distribution=$(. /etc/os-release;echo $ID$VERSION_ID) \
+        && curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey \
+            | sudo gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg \
+        && curl -s -L https://nvidia.github.io/libnvidia-container/$distribution/libnvidia-container.list \
+            | sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' \
+            | sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list \
+    && sudo apt update \
     && sudo apt install -y --no-install-recommends \
+        nvidia-container-toolkit \
+    && sudo systemctl restart docker
+fi
+
+# (command -v nvidia-docker >/dev/null 2>&1 || dpkg -l | grep -q nvidia-container-toolkit)
+
+# && sudo apt install -y --no-install-recommends \
+
